Reuse schema tests for inline validator regressions
